requirements.txt
selenium
pandas
openpyxl
google-cloud-bigquery
Flask
pytz
